address Phil's comments
