--- samples/SoundBoard/src/main/java/com/google/oboe/samples/soundboard/TileListener.java
@@ -1,7 +1,0 @@
-package com.google.oboe.samples.soundboard;
-
-public interface TileListener {
-    public void onTileOn(int index);
-    public void onTileOff(int index);
-}
-
